1.修改人员登陆注册返回结果 2.自行车数据表字段类型修改

+++ src/Java/com/yk/pojo/BikeInfo.java
@@ -6,7 +6,7 @@ public class BikeInfo {
     private String bikeId;
     private double latitude;
     private double longitude;
-    private int mileage;
+    private float mileage;
     private String fix;
 
     public int getId() {
@@ -54,11 +54,11 @@ public class BikeInfo {
         return this;
     }
 
-    public int getMileage() {
+    public float getMileage() {
         return mileage;
     }
 
-    public BikeInfo setMileage(int mileage) {
+    public BikeInfo setMileage(float mileage) {
         this.mileage = mileage;
         return this;
     }
